Solved issue #18
