+++ C++/Armstrong.cpp
@@ -1,0 +1,20 @@
+#include <iostream>  
+using namespace std;  
+int main()  
+{  
+int n,r,sum=0,temp;    
+cout<<"Enter the Number=  ";    
+cin>>n;    
+temp=n;    
+while(n>0)    
+{    
+r=n%10;    
+sum=sum+(r*r*r);    
+n=n/10;    
+}    
+if(temp==sum)    
+cout<<"Armstrong Number."<<endl;    
+else    
+cout<<"Not Armstrong Number."<<endl;   
+return 0;  
+}
