+++ C++/BinarytoDecimal.cpp
