+++ C++/binary_search.cpp
