+++ C/matrix_arithmetic.c
