+++ C/more simple c.c
