+++ C/palindromeChecker.c
