+++ C/structureusecase.c
